Compiling datapoint extension rewrite
